v5-to-do-list-drf-react/backend/requirements.txt
asgiref==3.8.1
bcrypt==4.2.0
Django==5.1
django-cors-headers==4.3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pyhumps==3.8.0
PyJWT==2.9.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
